less deprecation warning logs
